+++ cocos/audio/apple/AudioDecoder.mm
@@ -81,7 +81,7 @@ namespace cocos2d { namespace experimental {
             _outputFormat.mBitsPerChannel = 16;
             _outputFormat.mFormatFlags = kAudioFormatFlagsNativeEndian | kAudioFormatFlagIsPacked | kAudioFormatFlagIsSignedInteger;
 
-            _sampleRate = _outputFormat.mSampleRate;
+            _sampleRate = static_cast<uint32_t>(_outputFormat.mSampleRate);
             _channelCount = _outputFormat.mChannelsPerFrame;
             _bytesPerFrame = 2 * _outputFormat.mChannelsPerFrame;
